Move shouldFetch* fns to ucr util
