+++ test/actions/composite.test.js
@@ -5,13 +5,7 @@ import sinon from 'sinon'
 
 import * as historyUtil from '../../src/util/history'
 import * as filterActions from '../../src/actions/filters'
-import {
-  shouldFetchUcr,
-  shouldFetchSummaries,
-  shouldFetchNibrs,
-  updateApp,
-} from '../../src/actions/composite'
-import { nationalKey } from '../../src/util/usa'
+import { updateApp } from '../../src/actions/composite'
 
 const noop = () => {}
 
@@ -25,42 +19,6 @@ describe('composite actions', () => {
 
   afterEach(() => {
     sandbox.restore()
-  })
-
-  describe('shouldFetchUcr()', () => {
-    it('should return false for national key', () => {
-      const result = shouldFetchUcr({ place: nationalKey })
-      expect(result).toEqual(false)
-    })
-
-    it('should return true for valid state names', () => {
-      const result = shouldFetchUcr({ place: 'montana' })
-      expect(result).toEqual(true)
-    })
-  })
-
-  describe('shouldFetchSummaries()', () => {
-    it('should return undefined if crime or place are missing', () => {
-      const result = shouldFetchSummaries({ place: nationalKey })
-      expect(result).toEqual(undefined)
-    })
-
-    it('should return a truthy value if crime and place are present', () => {
-      const result = shouldFetchSummaries({ crime: 'homicide', place: 'montana' })
-      expect(!!result).toEqual(true)
-    })
-  })
-
-  describe('shouldFetchNibrs()', () => {
-    it('should return false if place does not submit NIBRS', () => {
-      const result = shouldFetchNibrs({ place: 'new-york' })
-      expect(result).toEqual(false)
-    })
-
-    it('should return an object with "initial-year"if place submits NIBRS', () => {
-      const result = shouldFetchNibrs({ place: 'montana' })
-      expect(Object.prototype.hasOwnProperty.call(result, 'initial-year')).toEqual(true)
-    })
   })
 
   describe('updateApp()', () => {
